Fix wrong type being used in password response
Can't take the address of a size_t and only write
8 bits out, even if the value is less than 255.

+++ src/ply-boot-server.c
@@ -198,7 +198,7 @@ ply_boot_connection_on_password_answer (ply_boot_connection_t *connection,
                                         ply_answer_t          *answer)
 {
 
-  size_t size;
+  uint8_t size;
 
   /* splash plugin isn't able to ask for password,
    * punt to client
